organize story files, add a11y test, autodoc stories
