+++ src/packages/svelte-lib/package.json
@@ -41,6 +41,7 @@
     "svelte": "3.54.0"
   },
   "devDependencies": {
+    "@storybook/addon-a11y": "7.0.23",
     "@storybook/addon-essentials": "7.1.0-alpha.39",
     "@storybook/addon-interactions": "7.1.0-alpha.39",
     "@storybook/addon-links": "7.1.0-alpha.39",
