+++ src/packages/svelte-lib/src/stories/Button/Button.stories.ts
@@ -6,7 +6,6 @@ import Button from "./Button.svelte";
 const meta = {
   title: "Example/Button",
   component: Button,
-  tags: ["autodocs"],
   argTypes: {
     backgroundColor: { control: "color" },
     size: {
