+++ src/packages/svelte-lib/src/stories/Header/Header.stories.ts
@@ -4,8 +4,6 @@ import Header from "./Header.svelte";
 const meta = {
   title: "Example/Header",
   component: Header,
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
-  tags: ["autodocs"],
   parameters: {
     // More on how to position stories at: https://storybook.js.org/docs/svelte/configure/story-layout
     layout: "fullscreen",
