+++ src/packages/svelte-lib/src/stories/Introduction.mdx
@@ -1,0 +1,7 @@
+import { Meta } from "@storybook/blocks";
+
+<Meta title="Introduction" />
+
+# Configure your project
+
+barbajoe's svelte library stories
