+++ src/packages/svelte-lib/src/stories/Page/Page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import './page.css';
-  import Header from './Header.svelte';
+  import "./page.css";
+  import Header from "../Header/Header.svelte";
 
   let user: { name: string } | null = null;
 </script>
@@ -8,9 +8,9 @@
 <article>
   <Header
     {user}
-    on:login={() => (user = { name: 'Jane Doe' })}
+    on:login={() => (user = { name: "Jane Doe" })}
     on:logout={() => (user = null)}
-    on:createAccount={() => (user = { name: 'Jane Doe' })}
+    on:createAccount={() => (user = { name: "Jane Doe" })}
   />
 
   <section class="storybook-page">
@@ -27,33 +27,46 @@
       process starting with atomic components and ending with pages.
     </p>
     <p>
-      Render pages with mock data. This makes it easy to build and review page states without
-      needing to navigate to them in your app. Here are some handy patterns for managing page data
-      in Storybook:
+      Render pages with mock data. This makes it easy to build and review page
+      states without needing to navigate to them in your app. Here are some
+      handy patterns for managing page data in Storybook:
     </p>
     <ul>
       <li>
-        Use a higher-level connected component. Storybook helps you compose such data from the
-        "args" of child component stories
+        Use a higher-level connected component. Storybook helps you compose such
+        data from the "args" of child component stories
       </li>
       <li>
-        Assemble data in the page component from your services. You can mock these services out
-        using Storybook.
+        Assemble data in the page component from your services. You can mock
+        these services out using Storybook.
       </li>
     </ul>
     <p>
       Get a guided tutorial on component-driven development at
-      <a href="https://storybook.js.org/tutorials/" target="_blank" rel="noopener noreferrer">
+      <a
+        href="https://storybook.js.org/tutorials/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
         Storybook tutorials
       </a>
       . Read more in the
-      <a href="https://storybook.js.org/docs" target="_blank" rel="noopener noreferrer">docs</a>
+      <a
+        href="https://storybook.js.org/docs"
+        target="_blank"
+        rel="noopener noreferrer">docs</a
+      >
       .
     </p>
     <div class="tip-wrapper">
       <span class="tip">Tip</span>
       Adjust the width of the canvas with the
-      <svg width="10" height="10" viewBox="0 0 12 12" xmlns="http://www.w3.org/2000/svg">
+      <svg
+        width="10"
+        height="10"
+        viewBox="0 0 12 12"
+        xmlns="http://www.w3.org/2000/svg"
+      >
         <g fill="none" fill-rule="evenodd">
           <path
             d="M1.5 5.2h4.8c.3 0 .5.2.5.4v5.1c-.1.2-.3.3-.4.3H1.4a.5.5 0
